+++ R/01_2_SANBA_aux_not_exported.R
@@ -1,34 +1,40 @@
-#' Compute posterior mean of stick-breaking weights.
+#' Compute Posterior Mean of Stick-Breaking Weights
 #'
-#' @param alk,blk the posterior variational parameters for the SB betas.
+#'
+#' @param alk the first posterior variational parameters for the stick-breaking beta variables.
+#' @param blk the second posterior variational parameters for the stick-breaking beta variables.
+#'
 #' @noRd
 #'
 post_sb_weight <- function(alk, blk){
-  
+
   K  <- ncol(alk)
   L  <- nrow(alk)
   # are we sure about this?
   logomega_post <- matrix(NA,L,K)
-  
+
   for(k in 1:K){
-    
+
     p2 <- blk[,k]/(alk[,k]+blk[,k])
     p2 <- c(1,p2[-L])
-    
+
     logomega_post[,k] <- log(alk[,k]/(alk[,k]+blk[,k])) + cumsum(log(p2))
-    
+
   }
-  
+
   return(exp(logomega_post))
-  
+
 }
 
 
 
-#' Relabel clusters 
+#' Relabeling Clusters
+#'
 #' @keywords internal
 #' @param ix vector of cluster labels
-.relabel <- function(ix) 
+#' @noRd
+#'
+.relabel <- function(ix)
 {
   if(min(ix)==0) ix <- ix+1
   while( max(ix) != length(unique(ix)) )
@@ -37,19 +43,5 @@ post_sb_weight <- function(alk, blk){
     ix[ix>missing_label1] <- ix[ix>missing_label1]-1
   }
   return(ix)
-}
-
-
-#' Extract best
-#' @param object 
-#' @keywords internal
-extract_best <- function(object) {
-
-  summa <- do.call(rbind,
-                   lapply(object,  function(q)
-                     c(max = max(q$sim$Elbo_val))))
-  ind <- which.max(summa)
-
-  return(object[[ind]])
 }